+++ music_videos/video_struct.go
@@ -1,0 +1,14 @@
+package music_videos
+
+import (
+	"fengqi/kodi-metadata-tmdb-cli/ffmpeg"
+)
+
+type MusicVideo struct {
+	Dir         string
+	Title       string
+	OriginTitle string
+	DateAdded   string
+	VideoStream *ffmpeg.Stream
+	AudioStream *ffmpeg.Stream
+}
